requirements.txt
anyio==3.6.2
click==8.0.3
colorama==0.4.5
fastapi==0.83.0
h11==0.13.0
idna==3.4
ImageHash==4.3.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector==2.2.9
numpy==1.19.5
Pillow==8.4.0
pydantic==1.9.2
python-multipart==0.0.5
PyWavelets==1.1.1
scipy==1.5.4
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.1.0
uvicorn==0.17.0